requirements.txt
streamlit
requests
PyMuPDF
pymupdf4llm
google-cloud-vision
google-cloud-aiplatform
futures
PyPDF2
python-docx
llama-index-core 
llama-parse
llama-index-readers-file 
setuptools
langfuse
sentry-sdk
pydantic_settings
pyyaml
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pandas
openpyxl
pytest
httpx<0.28